requirements.txt
opencv-python==4.8.0.76
pillow==9.5.0
tqdm==4.65.0
